+++ core/domain/src/main/java/com/abnamro/core/domain/usecase/GetRepoDetailsUsecase.kt
@@ -6,7 +6,7 @@ import com.abnamro.core.domain.repository.RepoRepository
 import javax.inject.Inject
 
 class GetRepoDetailsUsecase @Inject constructor(private val repository: RepoRepository) {
-    suspend operator fun invoke(id: Int): Result<RepoBO> {
+    suspend operator fun invoke(id: Long): Result<RepoBO> {
         return repository.getRepoDetail(id)
     }
 }
